tools/atlas-pipeline/requirements.txt
polars>=0.19
pydantic>=2.0
lxml>=4.9
beautifulsoup4>=4.12
requests>=2.31
mwparserfromhell>=0.6.4
PyYAML>=6.0
